rm duplicated code
